--- a/src/fields/fp.rs
+++ b/src/fields/fp.rs
@@ -411,6 +411,12 @@ impl Field for Fp {
         0x700e6467ac19ef1d,
         0x376bc3c62040b13f,
     ];
+    const BETA: Self = Fp::from_raw([
+        0x33fffcfa00000002,
+        0xc37d383464024905,
+        0x96d4c09e11330dde,
+        0x5c5e464a35c12768,
+    ]);
 
     fn is_zero(&self) -> Choice {
         self.ct_eq(&Self::zero())