@ebfull 2019-09-17
All of the further optimizations won't get us below 2^19 probably, so I'm not going to bother right now.